Relación jugador-partida implementada Fixes#80

+++ src/main/java/org/springframework/samples/petclinic/jugador/JugadorRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JugadorRepository extends CrudRepository<Jugador, Long> {
+public interface JugadorRepository extends CrudRepository<Jugador, Integer> {
 
     List<Jugador> findAll();
